# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: Invalid credentials or user not in allowed group
        type: string
    type: object
...
  models.GenericSuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  models.SSHAccessRequest:
    properties:
      duration:
        description: How long the access should be valid for (e.g., "1h", "30m")
        type: string
      sshUsername:
        description: Optional override for container's SSH user
        type: string
    type: object
  models.SSHAccessResponse:
    properties:
      command:
        description: Example SSH command
        type: string
      expiration:
        description: When this access expires
        type: string
      host:
        description: API server's hostname or IP
        type: string
      port:
        description: Allocated port on API server
        type: integer
      username:
        description: Username to use for SSH
        type: string
    type: object
  models.SSHSessionInfo:
    properties:
      created:
        description: When this access was created
        type: string
      expiration:
        description: When this access expires
        type: string
      labName:
        description: Lab name
        type: string
      nodeName:
        description: Node name
        type: string
      port:
        description: Allocated port on API server
        type: integer
      username:
        description: SSH username
        type: string
    type: object
...
  models.TerminalProtocol:
    enum:
    - ssh
    - shell
    - telnet
    type: string
    x-enum-varnames:
    - TerminalProtocolSSH
    - TerminalProtocolShell
    - TerminalProtocolTelnet
  models.TerminalSessionInfo:
    properties:
      createdAt:
        type: string
      error:
        type: string
      exitCode:
        type: integer
      expiresAt:
        type: string
      labName:
        type: string
      lastActivity:
        type: string
      nodeName:
        type: string
      protocol:
        $ref: '#/definitions/models.TerminalProtocol'
      sessionId:
        type: string
      state:
        type: string
      username:
        type: string
    type: object
  models.TerminalSessionRequest:
    properties:
      cols:
        description: Cols is the requested terminal width in characters.
        type: integer
      protocol:
        allOf:
        - $ref: '#/definitions/models.TerminalProtocol'
        description: 'Protocol selects the terminal transport: ssh, shell, or telnet.'
      rows:
        description: Rows is the requested terminal height in characters.
        type: integer
      sshUsername:
        description: SSHUsername optionally overrides the SSH username for protocol=ssh.
        type: string
      telnetPort:
        description: TelnetPort optionally overrides the telnet destination port for
          protocol=telnet.
        type: integer
    required:
    - protocol
    type: object
...
info:
  contact:
    name: Containerlab API Server maintainers
    url: https://github.com/srl-labs/clab-api-server/issues
  description: |-
    REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.

    Most endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer <token>.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Containerlab API Server
  version: "1.0"
paths:
...
  /api/v1/labs/{labName}/nodes/{nodeName}/ssh:
    post:
      consumes:
      - application/json
      description: |-
        Creates temporary SSH access to a lab node and returns connection details for an external SSH client.
        This endpoint does not create a browser terminal session. For browser terminals, use `/api/v1/labs/{labName}/nodes/{nodeName}/terminal-sessions` with `protocol=ssh`.
      parameters:
      - description: Lab name
        in: path
        name: labName
        required: true
        type: string
      - description: Full container name of the node (e.g., clab-my-lab-srl1)
        in: path
        name: nodeName
        required: true
        type: string
      - description: SSH access parameters
        in: body
        name: sshRequest
        schema:
          $ref: '#/definitions/models.SSHAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SSH connection details
          schema:
            $ref: '#/definitions/models.SSHAccessResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden (not owner of the lab)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Lab or node not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request temporary SSH access details
      tags:
      - SSH Access
...
  /api/v1/labs/{labName}/nodes/{nodeName}/terminal-sessions:
    post:
      consumes:
      - application/json
      description: |-
        Creates a constrained interactive terminal session for an owned lab node using one of the supported protocols: `ssh`, `shell`, or `telnet`.

        **Protocol behavior**
        - `ssh`: launches the server-side `ssh` client in a PTY directly to the node management IP and streams that terminal over WebSocket.
        - `shell`: launches a PTY-backed `docker|podman exec -it <container> <server-selected-command>` session.
        - `telnet`: launches a PTY-backed `docker|podman exec -it <container> telnet 127.0.0.1 <port>` session.

        **Security notes**
        - The backend chooses the launch command for every protocol. Clients choose only the protocol and terminal size.
        - Access is limited to owned nodes unless the caller is a configured superuser.

        **Relationship to `/api/v1/labs/{labName}/nodes/{nodeName}/ssh`**
        - This endpoint creates the interactive browser terminal session.
        - The separate `/ssh` endpoint returns temporary SSH access details for an external SSH client and is not required for browser terminals.
      parameters:
      - description: Lab name
        in: path
        name: labName
        required: true
        type: string
      - description: Full container name of the node (e.g., clab-my-lab-srl1)
        in: path
        name: nodeName
        required: true
        type: string
      - description: Terminal session creation parameters
        in: body
        name: terminal_request
        required: true
        schema:
          $ref: '#/definitions/models.TerminalSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created terminal session
          schema:
            $ref: '#/definitions/models.TerminalSessionInfo'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Lab or node not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Too many active terminal sessions
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create browser terminal session
      tags:
      - Terminal Sessions
...
  /api/v1/ssh/sessions:
    get:
      description: |-
        Returns active SSH sessions.

        **Notes**
        - Regular users see only their sessions.
        - Superusers can include all sessions via the `all` query parameter.
      parameters:
      - description: 'If true and user is superuser, shows sessions for all users
          (default: false)'
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of active SSH sessions
          schema:
            items:
              $ref: '#/definitions/models.SSHSessionInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden (non-superuser attempting to list all sessions)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List SSH sessions
      tags:
      - SSH Access
  /api/v1/ssh/sessions/{port}:
    delete:
      description: Terminates a specific SSH session by port.
      parameters:
      - description: SSH session port to terminate
        in: path
        name: port
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session terminated successfully
          schema:
            $ref: '#/definitions/models.GenericSuccessResponse'
        "400":
          description: Invalid port parameter
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden (not owner of the session)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Terminate SSH session
      tags:
      - SSH Access
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"